refactor(investment): 백테스트를 전략관리 페이지 내 인라인 뷰로 변경
- BacktestPanel 컴포넌트 생성 (기존 backtest/page.tsx 로직 분리)
- 전략관리 페이지에서 백테스팅 버튼 클릭 시 페이지 전환 없이
  동일 영역에 BacktestPanel 인라인 렌더링
- 뒤로 가기(←) 버튼으로 전략 목록으로 복귀
- 기존 /strategy/[id]/backtest URL 직접 접근 시 전략 목록으로 리다이렉트
